feat: Bulk invoice creation with duplicate check Fixes #1
This commit implements the functionality to create invoices in bulk by sending a list of invoices to the /bulk endpoint. Before saving the invoices, the system checks if each invoice already exists in the database. If any of the invoices in the list already exist, the entire operation is rolled back, and an error message is displayed, indicating which invoices already exist. If all the invoices in the list are new, they will be saved to the database. Any new invoices that do not already exist will be successfully saved.